Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-    int a,b;
-    scanf("%d%d",&a,&b);
-    printf("Product: %d",a*b);
+    float a,b;
+    scanf("%f%f",&a,&b);
+    printf("Product: %.2f",a*b);
 }
